server/requirements.txt
pydantic>=2.5
fastapi>=0.110
uvicorn[standard]>=0.29
requests>=2.31
supabase>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0